Convert SLC factor as absolute value.

+++ sbk-api/src/main/java/io/sbk/logger/impl/GrpcPrometheusLogger.java
@@ -252,7 +252,7 @@ public class GrpcPrometheusLogger extends PrometheusLogger {
     @Override
     public void print(double seconds, long bytes, long records, double recsPerSec, double mbPerSec, double avgLatency,
                       long maxLatency, long invalid, long lowerDiscard, long higherDiscard,
-                      int slc1, int slc2, long[] percentileValues) {
+                      long slc1, long slc2, long[] percentileValues) {
         super.print(seconds, bytes, records, recsPerSec, mbPerSec, avgLatency, maxLatency, invalid, lowerDiscard,
                 higherDiscard, slc1, slc2, percentileValues);
         if (latencyBytes > 0 ) {
