+++ sbk-api/src/main/java/io/sbk/logger/impl/PrometheusLogger.java
@@ -141,7 +141,7 @@ public class PrometheusLogger extends CSVLogger {
 
     private void printMetrics(double seconds, long bytes, long records, double recsPerSec, double mbPerSec,
                               double avgLatency, long maxLatency, long invalid, long lowerDiscard, long higherDiscard,
-                              int slc1, int slc2, long[] percentileValues) {
+                              long slc1, long slc2, long[] percentileValues) {
         super.print(seconds, bytes, records, recsPerSec, mbPerSec, avgLatency, maxLatency,
                 invalid, lowerDiscard, higherDiscard, slc1, slc2, percentileValues);
         prometheusServer.print(seconds, bytes, records, recsPerSec, mbPerSec, avgLatency, maxLatency,
@@ -151,7 +151,7 @@ public class PrometheusLogger extends CSVLogger {
     @Override
     public void print(double seconds, long bytes, long records, double recsPerSec, double mbPerSec, double avgLatency,
                       long maxLatency, long invalid, long lowerDiscard, long higherDiscard,
-                      int slc1, int slc2, long[] percentileValues) {
+                      long slc1, long slc2, long[] percentileValues) {
         printer.print(seconds, bytes, records, recsPerSec, mbPerSec, avgLatency, maxLatency,
                 invalid, lowerDiscard, higherDiscard, slc1, slc2,  percentileValues);
     }
