+++ sbk-api/src/main/java/io/sbk/perl/Print.java
@@ -23,11 +23,11 @@ public interface Print {
      * @param invalid Number of invalid/negative latencies.
      * @param lowerDiscard number of discarded latencies which are less than minimum latency.
      * @param higherDiscard number of discarded latencies which are higher than maximum latency.
-     * @param slc1  Sliding Latency Coverage percentage
-     * @param slc2  Sliding Latency Coverage percentage
+     * @param slc1  Sliding Latency Coverage factor
+     * @param slc2  Sliding Latency Coverage factor
      * @param percentiles Array of percentiles.
      */
     void print(double seconds, long bytes, long records, double recsPerSec, double mbPerSec, double avgLatency,
                long maxLatency, long invalid, long lowerDiscard, long higherDiscard,
-               int slc1, int slc2, long[] percentiles);
+               long slc1, long slc2, long[] percentiles);
 }
